+++ hammr/commands/image/image.py
@@ -28,7 +28,8 @@ from ussclicore.utils import generics_utils, printer, progressbar_widget, downlo
 from hammr.utils import *
 from hammr.utils.hammr_utils import *
 from hammr.utils.deployment_utils import *
-from uforge.objects.uforge import *
+from hammr.utils.publish_utils import *
+from hammr.utils.publish_builders import *
 
 #This import and configuration avoid pyxb warnings about xmls
 import logging
@@ -86,7 +87,7 @@ class Image(Cmd, CoreGlobal):
             table.header(["Template name", "Image ID", "Publish ID", "Account name", "Format", "Cloud ID", "Status"])
             publish_images = generics_utils.order_list_object_by(publish_images, "name")
             for publish_image in publish_images:
-                pubStatus = self.get_publish_status(publish_image.status)
+                pubStatus = get_publish_status(publish_image.status)
                 table.add_row([publish_image.name,
                                generics_utils.extract_id(publish_image.imageUri),
                                publish_image.dbId,
@@ -98,7 +99,7 @@ class Image(Cmd, CoreGlobal):
 
             return 0
         except ArgumentParserError as e:
-            printer.out("ERROR: In Arguments: " + str(e), printer.ERROR)
+            printer.out("In Arguments: " + str(e), printer.ERROR)
             self.help_list()
         except Exception as e:
             return handle_uforge_exception(e)
@@ -108,90 +109,36 @@ class Image(Cmd, CoreGlobal):
         doParser.print_help()
 
     def arg_publish(self):
-        doParser = ArgumentParser(prog=self.cmd_name + " publish", add_help=True,
+        do_parser = ArgumentParser(prog=self.cmd_name + " publish", add_help=True,
                                   description="Publish (upload and register) a built machine image to a target environment")
-        mandatory = doParser.add_argument_group("mandatory arguments")
+        mandatory = do_parser.add_argument_group("mandatory arguments")
         mandatory.add_argument('--file', dest='file', required=True,
                                help="yaml/json file providing the cloud account parameters required for upload and registration")
-        optional = doParser.add_argument_group("optional arguments")
+        optional = do_parser.add_argument_group("optional arguments")
         optional.add_argument('--id', dest='id', required=False, help="id of the image to publish")
-        return doParser
+        return do_parser
 
     def do_publish(self, args):
         try:
-            # add arguments
-            doParser = self.arg_publish()
-            doArgs = doParser.parse_args(shlex.split(args))
+            do_args = self.parse_args(args)
+            template = retrieve_template_from_file(do_args.file)
+            if do_args.id:
+                self.do_publish_with_id(do_args.id, template)
+            else:
+                self.do_publish_without_id(template)
 
-            #if the help command is called, parse_args returns None object
-            if not doArgs:
-                    return 2
-
-            file = generics_utils.get_file(doArgs.file)
-            if file is None:
-                return 2
-            template = validate(file)
-            if template is None:
-                return
-
-            try:
-                if doArgs.id:
-                    images = self.get_all_images()
-                    image = self.get_image(images, str(doArgs.id))
-                    if image is None:
-                        printer.out("Image not found", printer.ERROR)
-                        return 2
-                    if not self.is_image_ready_to_publish(image, None):
-                        printer.out("Image with name '" + image.name + " can not be published", printer.ERROR)
-                        return 2
-                    appliance = self.api.Users(self.login).Appliances(
-                        generics_utils.extract_id(image.applianceUri)).Get()
-                    if appliance is None or not hasattr(appliance, 'dbId'):
-                        printer.out("No template found for image", printer.ERROR)
-                        return
-                    rInstallProfile = self.api.Users(self.login).Appliances(appliance.dbId).Installprofile("").Get()
-                    if rInstallProfile is None:
-                        printer.out("No installation found on the template '" + template["stack"]["name"] + "'",
-                                    printer.ERROR)
-                        return
-                    builder = self.find_builder(image, template)
-                    if builder is None:
-                        # TODO unmap image format
-                        printer.out("No builder part found for image with format type: " + str(template["type"]),
-                                    printer.ERROR)
-                        return 2
-                    self.publish_builder(builder, template, appliance, rInstallProfile, 1, image)
-                else:
-                    # Get template which correpond to the template file
-                    appliances = self.api.Users(self.login).Appliances().Getall(
-                        Query="name=='" + template["stack"]["name"] + "';version=='" + template["stack"][
-                            "version"] + "'")
-                    appliance = appliances.appliances.appliance
-                    if appliance is None or len(appliance) != 1:
-                        printer.out("No template found on the plateform", printer.ERROR)
-                        return 0
-                    appliance = appliance[0]
-                    rInstallProfile = self.api.Users(self.login).Appliances(appliance.dbId).Installprofile("").Get()
-                    if rInstallProfile is None:
-                        printer.out("No installation found on the template '" + template["stack"]["name"] + "'",
-                                    printer.ERROR)
-                        return
-
-                    i = 1
-                    for builder in template["builders"]:
-                        rCode = self.publish_builder(builder, template, appliance, rInstallProfile, i, None)
-                        if rCode >= 2:
-                            return
-                        i += 1
-
-            except KeyError as e:
-                printer.out("unknown error template file, key: " + str(e), printer.ERROR)
-
+        except KeyError as e:
+            printer.out("unknown error template file, key: " + str(e), printer.ERROR)
+            return 2
         except ArgumentParserError as e:
-            printer.out("ERROR: In Arguments: " + str(e), printer.ERROR)
+            printer.out("In Arguments: " + str(e), printer.ERROR)
+            return 2
             self.help_publish()
+        except ValueError as e:
+            printer.out(str(e), printer.ERROR)
+            return 2
         except KeyboardInterrupt:
-            pass
+            return 2
         except Exception as e:
             return handle_uforge_exception(e)
 
@@ -251,7 +198,7 @@ class Image(Cmd, CoreGlobal):
             return 2
 
         except ArgumentParserError as e:
-            printer.out("ERROR: In Arguments: " + str(e), printer.ERROR)
+            printer.out("In Arguments: " + str(e), printer.ERROR)
             self.help_deploy()
         except KeyboardInterrupt:
             printer.out(
@@ -308,7 +255,7 @@ class Image(Cmd, CoreGlobal):
                 printer.out("Image not found", printer.ERROR)
 
         except ArgumentParserError as e:
-            printer.out("ERROR: In Arguments: " + str(e), printer.ERROR)
+            printer.out("In Arguments: " + str(e), printer.ERROR)
             self.help_delete()
             return 2
         except ValueError as e:
@@ -365,7 +312,7 @@ class Image(Cmd, CoreGlobal):
                 printer.out("Image not found", printer.ERROR)
 
         except ArgumentParserError as e:
-            printer.out("ERROR: In Arguments: " + str(e), printer.ERROR)
+            printer.out("In Arguments: " + str(e), printer.ERROR)
             self.help_delete()
             return 2
         except ValueError as e:
@@ -426,7 +373,7 @@ class Image(Cmd, CoreGlobal):
 
 
         except ArgumentParserError as e:
-            printer.out("ERROR: In Arguments: " + str(e), printer.ERROR)
+            printer.out("In Arguments: " + str(e), printer.ERROR)
             self.help_download()
             return 2
         except ValueError as e:
@@ -439,6 +386,13 @@ class Image(Cmd, CoreGlobal):
         doParser = self.arg_download()
         doParser.print_help()
 
+    def parse_args(self, args):
+        do_parser = self.arg_publish()
+        do_args = do_parser.parse_args(shlex.split(args))
+        if not do_args:
+            raise ArgumentParserError("No arguments")
+        return do_args
+
     def get_image_status(self, status):
         if (status.complete and not status.error):
             imgStatus = "Done"
@@ -450,101 +404,42 @@ class Image(Cmd, CoreGlobal):
             imgStatus = "In progress (" + str(status.percentage) + "%)"
         return imgStatus
 
-    def get_publish_status(self, status):
-        if (status.complete and not status.error):
-            pubStatus = "Done"
-        elif status.error:
-            pubStatus = "Error"
-        elif status.cancelled:
-            pubStatus = "Canceled"
+    def do_publish_with_id(self, id, template):
+        images = self.get_all_images()
+        image = self.get_image(images, str(id))
+        if image is None:
+            raise ValueError("Image not found")
+        if not is_image_ready_to_publish(image, None):
+            raise ValueError("Image with name '" + image.name + "' can not be published")
+
+        source = retrieve_source_from_image(self, image)
+        builder = self.find_builder(image, template)
+        publish_image_from_builder(self, builder, template, source, 1, image)
+
+    def do_publish_without_id(self, template):
+        if template.has_key("stack") and template["stack"].has_key("name") and template["stack"].has_key("version"):
+            query_string = "name=='" + template["stack"]["name"] + "';version=='" + template["stack"]["version"] + "'"
+            appliance = self.retrieve_appliances_from_name_and_version(query_string)
+            appliance = appliance[0]
+            publish_all_builders(self, template, appliance)
+
+    def retrieve_publish_image_with_target_format_builder(self, image, builder):
+        format_type = image.targetFormat.format.name
+        publish_method = getattr(publish_builders, "publish_" + generics_utils.remove_special_chars(format_type), None)
+        if publish_method:
+            publish_image = publish_method(builder)
+            if publish_image is None:
+                raise ValueError("Could not find the builder")
         else:
-            pubStatus = "In progress (" + str(status.percentage) + "%)"
-        return pubStatus
+            raise ValueError("Builder type unknown: " + format_type)
+        return publish_image
 
-    def publish_builder(self, builder, template, appliance, rInstallProfile, i, comliantImage):
-        try:
-            if comliantImage is None:
-                comliantImage = self.get_image_to_publish(builder, template, appliance, i)
-
-            # get target format to define publish method
-            format_type = comliantImage.targetFormat.format.name
-            publishMethod = getattr(publish_utils, "publish_" + generics_utils.remove_special_chars(format_type), None)
-            if publishMethod:
-                mypImage = publishMethod(builder)
-                if mypImage is None:
-                    return 2
-            else:
-                printer.out("Builder type unknown: " + format_type, printer.ERROR)
-                return 2
-
-            mypImage.imageUri = comliantImage.uri
-            mypImage.applianceUri = appliance.uri
-            mypImage.credAccount = self.get_account_to_publish(builder)
-            account_name = mypImage.credAccount.name
-
-            rpImage = self.api.Users(self.login).Appliances(appliance.dbId).Images(
-                comliantImage.dbId).Pimages().Publish(body=mypImage, element_name="ns1:publishImage")
-
-            status = rpImage.status
-            statusWidget = progressbar_widget.Status()
-            statusWidget.status = status
-            widgets = [Bar('>'), ' ', statusWidget, ' ', ReverseBar('<')]
-            progress = ProgressBar(widgets=widgets, maxval=100).start()
-            while not (status.complete or status.error or status.cancelled):
-                statusWidget.status = status
-                progress.update(status.percentage)
-                status = self.api.Users(self.login).Appliances(appliance.dbId).Images(comliantImage.dbId).Pimages(
-                    rpImage.dbId).Status.Get()
-                time.sleep(2)
-            statusWidget.status = status
-            progress.finish()
-            if status.error:
-                printer.out("Publication to '" + builder["account"][
-                    "name"] + "' error: " + status.message + "\n" + status.errorMessage, printer.ERROR)
-                if status.detailedError:
-                    printer.out(status.detailedErrorMsg)
-            elif status.cancelled:
-                printer.out("\nPublication to '" + builder["account"][
-                    "name"] + "' canceled: " + status.message.printer.WARNING)
-            else:
-                printer.out("Publication to " + account_name + " is ok", printer.OK)
-                rpImage = self.api.Users(self.login).Appliances(appliance.dbId).Images(comliantImage.dbId).Pimages(
-                    rpImage.dbId).Get()
-                if rpImage.cloudId is not None and rpImage.cloudId != "":
-                    printer.out("Cloud ID : " + rpImage.cloudId)
-            return 0
-        except KeyboardInterrupt:
-            printer.out("\n")
-            if generics_utils.query_yes_no("Do you want to cancel the job ?"):
-                if 'appliance' in locals() and 'comliantImage' in locals() and 'rpImage' in locals() \
-                        and hasattr(appliance, 'dbId') and hasattr(comliantImage, 'dbId') and hasattr(rpImage, 'dbId'):
-                    self.api.Users(self.login).Appliances(appliance.dbId).Images(comliantImage.dbId).Pimages(
-                        rpImage.dbId).Cancel.Cancel()
-                else:
-                    printer.out("Impossible to cancel", printer.WARNING)
-            else:
-                printer.out("Exiting command")
-            raise KeyboardInterrupt
-
-    def is_image_ready_to_publish(self, image, builder):
-        if builder is not None:
-            if ("hardwareSettings" in builder) and ("memory" in builder["hardwareSettings"]) and (
-                    not image.installProfile.memorySize == builder["hardwareSettings"]["memory"]):
-                return False
-
-            if ("installation" in builder) and ("swapSize" in builder["installation"]) and (
-                    not image.installProfile.swapSize == builder["installation"]["swapSize"]):
-                return False
-                # print str(image.installProfile.swapSize)+"----"+ str(builder["installation"]["swapSize"])
-                # TODO
-                # if ("diskSize" in builder["installation"]) and (not image.installProfile.diskSize == builder["installation"]["diskSize"]):
-                #        isOk = False
-                #        print str(image.installProfile.diskSize)+"----"+ str(builder["installation"]["diskSize"])
-
-        if not image.status.complete or image.status.error or image.status.cancelled:
-            return False
-
-        return True
+    def retrieve_appliances_from_name_and_version(self, query_string):
+        appliances = self.api.Users(self.login).Appliances().Getall(Query=query_string)
+        appliance = appliances.appliances.appliance
+        if appliance is None or len(appliance) != 1:
+            raise ValueError("No template found on the platform")
+        return appliance
 
     def is_publish_image_ready_to_deploy(self, publishImage):
         if not publishImage.status.complete or publishImage.status.error or publishImage.status.cancelled:
@@ -552,95 +447,11 @@ class Image(Cmd, CoreGlobal):
 
         return True
 
-
     def find_builder(self, image, template):
         for builder in template["builders"]:
             if image.targetFormat.name == builder["type"]:
                 return builder
-
-        return None
-
-    def get_image_to_publish(self, builder, template, appliance, i):
-        printer.out(
-            "Publishing '" + builder["type"] + "' image (" + str(i) + "/" + str(len(template["builders"])) + ")")
-        images = self.api.Users(self.login).Appliances(appliance.dbId).Images.Getall(
-            Query="targetFormat.name=='" + builder["type"] + "'")
-        images = images.images.image
-        if images is None or len(images) == 0:
-            printer.out(
-                "No images found for the template '" + template["stack"]["name"] + "' with type: " + builder[
-                    "type"], printer.ERROR)
-            return 2
-
-        compliantImages = []
-        for image in images:
-            isOk = self.is_image_ready_to_publish(image, builder)
-            if isOk:
-                compliantImages.append(image)
-
-        if (len(compliantImages) == 0):
-            printer.out(
-                "No images found for the template '" + template["stack"]["name"] + "' with type: " + builder[
-                    "type"], printer.ERROR)
-            return 2
-        elif (len(compliantImages) == 1):
-            comliantImage = compliantImages[0]
-        else:
-            # TODO get the last created image
-            comliantImage = compliantImages[0]
-
-        return comliantImage
-
-    def get_account_to_publish(self, builder):
-        if not "account" in builder:
-            printer.out("Missing account section on builder: [" + builder["type"] + "]", printer.ERROR)
-            return 2
-        # Get all cloud account on the plateform (for the user)
-        accounts = self.api.Users(self.login).Accounts.Getall()
-        accounts = accounts.credAccounts.credAccount
-        if accounts is None or len(accounts) == 0:
-            printer.out("No accounts available on the plateform.\n You can use the command 'hammr account create' to create an account.", printer.ERROR)
-            return 2
-        else:
-            for account in accounts:
-
-                builder_name = self.get_account_name_for_publish(builder, account)
-                if builder_name == "":
-                    printer.out("No account name given", printer.ERROR)
-                    return 2
-
-                if account.name == builder_name:
-                    # A hack to avoid a toDOM, toXML bug
-                    account.targetPlatform.type._ExpandedName = pyxb.namespace.ExpandedName(Namespace, 'string')
-                    if hasattr(account, 'certificates'):
-                        account.certificates._ExpandedName = pyxb.namespace.ExpandedName(Namespace, 'Certificates')
-                        for c in account.certificates.certificate:
-                            c.type._ExpandedName = pyxb.namespace.ExpandedName(Namespace, 'string')
-
-                    return account
-                    break
-
-            printer.out("No accounts available with name " + builder["account"]["name"] + ".\nYou can use the command 'hammr account create' to create an account.", printer.ERROR)
-            return 2
-
-    # get the account name from field or from the credential file given in the builder
-    def get_account_name_for_publish(self, builder, account):
-        builder_name = ""
-        if builder["account"].has_key("name"):
-            builder_name = builder["account"]["name"]
-
-        elif builder["account"].has_key("file"):
-
-            file = generics_utils.get_file(builder["account"]["file"])
-            if file is None:
-                return ""
-            template = validate(file)
-            if template is None:
-                return ""
-
-            builder_name = self.get_account_name_from_template(template, builder)
-
-        return builder_name
+        raise ValueError("No builder part found for image with this format type")
 
     def get_account_name_from_template(self, template, builder):
         account_name = ""