+++ hammr/utils/hammr_utils.py
@@ -28,6 +28,7 @@ import pyxb
 from uforge.objects.uforge import *
 import ussclicore.utils.download_utils
 from ussclicore.utils import download_utils
+from ussclicore.argumentParser import ArgumentParserError
 from ussclicore.utils import printer
 from ussclicore.utils import generics_utils
 from hammr.utils.bundle_utils import *
@@ -280,6 +281,16 @@ def extract_appliance_id(image_uri):
         return int(match.group(1))
     else:
         return None
+
+
+def retrieve_template_from_file(file):
+    file = generics_utils.get_file(file)
+    if not file:
+        raise ArgumentParserError("Wrong file argument")
+    template = validate(file)
+    if not template:
+        raise ValueError("Could not extract information from file")
+    return template
 
 def download_binary_in_local_temp_dir(api, local_temp_dir, uri_binary, binary_name):
     uri = generics_utils.get_uforge_url_from_ws_url(api.getUrl())