+++ documentation/en/source/changelog.rst
@@ -3,6 +3,15 @@
 Changelog
 =========
 
+hammr 3.8.3 (2018-04-16)
+------------------------
+
+Evolutions:
+
+* hammr image publish now works for images coming from scans
+* Remove End-Of-Life Microsoft Azure Classic
+* Add support for UForge API Keys
+
 hammr 3.8.2 (2018-03-05)
 ------------------------